pyproject.toml
[tool.poetry]
name = "dusty-bot-v2"
version = "0.1.0"
description = ""
authors = ["Ky Nguyen <ngky517@gmail.com>"]
packages = [
    { include = "services" }
]

[tool.poetry.dependencies]
python = "^3.10.5"
gunicorn = "^20.1"
flasgger = "=0.9.4dev0"
requests = "webargs"
"discord.py" = "^2.2.3"
webargs = "^8.1.0"
psycopg2-binary = "^2.9.3"
asyncpg = "^0.27.0"
aiohttp = "^3.8.1"
Flask = "^2.3.2"
Flask-SQLAlchemy = "^3.0.3"
Flask-Migrate = "^4.0.4"

[tool.poetry.dev-dependencies]
pytest = "^7.0.1"
pytest-random-order = "^1.0.4"
pylint = "~2.13.9"
pytest-cov = "^3.0"
responses = "^0.10.8"
honcho = "^1.1.0"
factory-boy = "^3.2.1"
"testing.postgresql" = "^1.3.0"
python-dotenv = "^1.0.0"

[tool.poetry.extras]
local_testing = ["testing.postgresql"]

[build-system]
requires = ["poetry>=1.0.3"]
build-backend = "poetry.masonry.api"
